pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jsleak"
version = "0.5.1"
description = "A production-quality tool to scan JavaScript files for exposed secrets and endpoints."
readme = "README.md"
authors = [
  { name = "Zain Nadeem", email = "zainnadeemzainnadeem80@gmail.com" },
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
]
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0"
]

[project.scripts]
jsleak = "jsleak.cli:main"

[project.urls]
Homepage = "https://github.com/zainnadeem786/jsleak"
Bug_Tracker = "https://github.com/zainnadeem786/jsleak/issues"
